+++ src/PhpCsFixer/Fixer/ImportFqcnInStaticCallFixer.php
@@ -40,7 +40,8 @@ use function in_array;
 use function mb_strtolower;
 
 /**
- * Import FQCNs used in static calls: \Vendor\Pkg\Cls::method() -> Cls::method() + use Vendor\Pkg\Cls;
+ * Import FQCNs used in static calls: \Vendor\Pkg\Cls::method() -> Cls::method()
+ * + use Vendor\Pkg\Cls;
  *
  * @author Brian Faust <brian@cline.sh>
  * @version 1.0.1